ques

+++ arraysLevel5/Ques8.java
@@ -3,12 +3,12 @@ public class Ques8 {
     public static void main(String[] args) {
         int arr[]={2,4,5,6,7};
         int sum=0;
-        int avg=0;
+        double avg=0;
         int count=0;
         for(int i=0;i<arr.length;i++){
             sum+=arr[i];
         }
-        avg=sum/arr.length;
+        avg=(double)sum/arr.length;
 
         for(int i=0;i<arr.length;i++){
             if(arr[i]>avg){
